とりあえずFirebaseのRealtime Databaseから、 child（1ユーザー）ごとのレコードをMap形式で取得することができた。
contextに加工して保持してるのを利用して、なんやかんやしていこう。

+++ app/src/main/java/com/example/testtwitter4j/bean/OutlayBean.kt
@@ -7,5 +7,5 @@ data class OutlayBean (
     var userId: String = "", // TwitterのID（@ 以降の）
     var addedDate: Date = Date(),
     var category: String = "",
-    var amount: Int = 0
+    var amount: Long = 0
 )
